Turn on stricter compiler warnings
And fix subsequent warnings caused by:

- shadowed variables (i.e. variables existing in nested scopes);

- signed vs. unsigned comparisons

- string pointers and buffers being unsigned which don't need to be;

- unnecessary casts;

- unused variables (or only used when debugging is enabled);

+++ libtac/lib/acct_r.c
@@ -98,7 +98,7 @@ int tac_acct_read(int fd, struct areply *re) {
     }
 
     spacket_read = read(fd, tb, ulen_from_header);
-    if(spacket_read < ulen_from_header) {
+    if(spacket_read < (ssize_t) ulen_from_header) {
         TACSYSLOG((LOG_ERR,\
             "%s: short reply body, read %zd of %zu: %m",\
             __FUNCTION__,\
